+++ src/api/apiNews.js
@@ -37,3 +37,16 @@ export const getCategories = async () => {
     console.log(error);
   }
 };
+
+export const getLatestNews = async () => {
+  try {
+    const response = await axios.get(`${BASE_URL}latest-news`, {
+      params: {
+        apiKey: API_KEY,
+      },
+    });
+    return response.data;
+  } catch (error) {
+    console.log(error);
+  }
+};
