+++ src/components/LatestNews/LatestNews.jsx
@@ -1,10 +1,13 @@
 import BannersList from "../BanersList/BannersList.jsx";
 import styles from "./styles.module.css";
+import { useFetch } from "../../helpers/hooks/useFetch.js";
+import { getLatestNews } from "../../api/apiNews.js";
 
-const LatestNews = ({ banners, isLoading }) => {
+const LatestNews = () => {
+  const { data, isLoading } = useFetch(getLatestNews, {});
   return (
     <section className={styles.section}>
-      <BannersList banners={banners} isLoading={isLoading} />
+      <BannersList banners={data && data.news} isLoading={isLoading} />
     </section>
   );
 };
