+++ src/components/NewsByFilters/NewsByFilters.jsx
@@ -1,10 +1,28 @@
-import { TOTAL_PAGES } from "../../constants/constants";
+import { PAGE_SIZE, TOTAL_PAGES } from "../../constants/constants";
 import NewsFilters from "../NewsFilters/NewsFilters";
 import NewsList from "../NewsList/NewsList";
-import Pagination from "../Pagination/Pagination";
 import styles from "./styles.module.css";
+import { useFilters } from "../../helpers/hooks/useFilters.js";
+import { useDebounce } from "../../helpers/hooks/useDebounce.js";
+import { useFetch } from "../../helpers/hooks/useFetch.js";
+import { getNews } from "../../api/apiNews.js";
+import PaginationWrapper from "../PaginationWrapper/PaginationWrapper.jsx";
 
-const NewsByFilters = ({ filters, changeFilter, isLoading, news }) => {
+const NewsByFilters = () => {
+  const { filters, changeFilter } = useFilters({
+    page_number: 1,
+    page_size: PAGE_SIZE,
+    category: null,
+    keywords: "",
+  });
+
+  const debouncedKeywords = useDebounce(filters.keywords, 1500);
+
+  const { data, isLoading } = useFetch(getNews, {
+    ...filters,
+    keywords: debouncedKeywords,
+  });
+
   const handleNextPage = () => {
     if (filters.page_number < TOTAL_PAGES) {
       changeFilter("page_number", filters.page_number + 1);
@@ -24,24 +42,17 @@ const NewsByFilters = ({ filters, changeFilter, isLoading, news }) => {
   return (
     <section className={styles.section}>
       <NewsFilters changeFilter={changeFilter} filters={filters} />
-
-      <Pagination
+      <PaginationWrapper
+        top
+        bottom
         handlePreviousPage={handlePreviousPage}
         handleNextPage={handleNextPage}
         handlePageClick={handlePageClick}
         totalPages={TOTAL_PAGES}
         currentPage={filters.page_number}
-      />
-
-      <NewsList isLoading={isLoading} news={news} />
-
-      <Pagination
-        handlePreviousPage={handlePreviousPage}
-        handleNextPage={handleNextPage}
-        handlePageClick={handlePageClick}
-        totalPages={TOTAL_PAGES}
-        currentPage={filters.page_number}
-      />
+      >
+        <NewsList isLoading={isLoading} news={data?.news} />
+      </PaginationWrapper>
     </section>
   );
 };
