+++ src/components/PaginationWrapper/styles.module.css
@@ -1,0 +1,1 @@
+/*No styles*/
